Fixed Vector3 and Quaternion definitions (moved to own header files)
